added static_cast<long int> to avoid an error which occurs when try to compile on RH7.1's gcc
git-svn-id: https://ecell.svn.sourceforge.net/svnroot/ecell/ecell3/trunk@176 f1531174-cb10-0410-9fe6-89aa7ac3eedb

+++ ecell/pyecs/PySimulator.cpp
@@ -148,7 +148,7 @@ Object PySimulator::getProperty( const Tuple& args )
 	}
       else if( aUniversalVariable.isInt() )
 	{
-	  anObject = Py::Int( aUniversalVariable.asInt() );
+	  anObject = Py::Int( static_cast<long int>( aUniversalVariable.asInt() ) );
 	}
       else if( aUniversalVariable.isString() )
 	{
